requirement.txt
Flask
mysql-connector-python
Werkzeug
APScheduler
python-dotenv